refactor: route ui imports through app facade
